src/tpm2-crypto.c: add call to Esys_Startup()
Esys_Startup() is needed for simulated TPMs. It doesn't result in error
for real TPM and does not reset PCR values, but this needs to be tested
on other vendors. According to TPM specification, this function should
result in TPM Reset when called without previous Shutdown(), although
observed results may suggest otherwise.

This is just for convenience when working on simulated TPM, in case of
any issues this function can be safely removed and 'tpm2_startup -c'
from tpm2-tool should be used instead (only once after each start of
simulation).

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>